fix: dispose in StatelessWidget removed, audio_joiner brace, pdf_splitter format, GeminiService.chat, GradientButton onPressed

+++ lib/screens/password_strength_screen.dart
@@ -237,7 +237,7 @@ class _PasswordAnalysis {
 
     return _PasswordAnalysis(
       score: score, label: labels[score], color: colors[score],
-      entropy: entropy, crackTime: crackTime, checks: checks, suggestions: suggestions,
+      entropy: entropy.toDouble(), crackTime: crackTime, checks: checks, suggestions: suggestions,
     );
   }
 }
